Update PS2 cover images in README for better display
The README file has been updated to adjust the size of the PlayStation 2 cover images. This change replaces the width attribute with a height attribute, setting each image to a consistent height of 200 pixels. Additionally, some new covers were added from the papx90516-pbpx95203 branch. This enhances the visual uniformity of the images on the page.

Release-Note: Adjust PS2 cover display in README for visual consistency

Signed-off-by: Ian Skelskey <ianskelskey@gmail.com>